refactor: ProfileFragment transported to the view package, toProfile() renamed to toProfileScreen()
