+++ app/src/main/java/com/demal/navigation/MainActivityNavigator.kt
@@ -1,5 +1,5 @@
 package com.demal.navigation
 
 interface MainActivityNavigator {
-    fun toProfile() //TODO: Change to navigation to first fragment
+    fun toProfileScreen() //TODO: Change to navigation to first fragment
 }
