+++ app/src/main/java/com/demal/navigation/Navigator.kt
@@ -1,6 +1,6 @@
 package com.demal.navigation
 
-import com.example.feat_profile.ProfileNavigator
+import com.example.feat_profile.navigation.ProfileNavigator
 import com.github.terrakok.cicerone.Router
 
 //Класс должен реализовать навигационные интерфейсы всех модулей
@@ -9,5 +9,5 @@ class Navigator(
     private val screens: Screens
 ) : MainActivityNavigator, ProfileNavigator {
 
-    override fun toProfile() = router.navigateTo(screens.profileScreen())
+    override fun toProfileScreen() = router.navigateTo(screens.profileScreen())
 }
