+++ app/src/main/java/com/demal/navigation/Screens.kt
@@ -1,6 +1,6 @@
 package com.demal.navigation
 
-import com.example.feat_profile.ProfileFragment
+import com.example.feat_profile.view.ProfileFragment
 import com.github.terrakok.cicerone.androidx.FragmentScreen
 
 //Все фрагменты должны быть объявленны здесь в виде FragmentScreen
