+++ feat_profile/src/main/java/com/example/feat_profile/navigation/ProfileNavigator.kt
@@ -1,4 +1,4 @@
-package com.example.feat_profile
+package com.example.feat_profile.navigation
 
 interface ProfileNavigator {
     //Методы для навигации ИЗ ProfileFragment
