+++ feat_profile/src/main/java/com/example/feat_profile/view/ProfileFragment.kt
@@ -1,6 +1,7 @@
-package com.example.feat_profile
+package com.example.feat_profile.view
 
 import androidx.fragment.app.Fragment
+import com.example.feat_profile.R
 
 //Только для теста навигации
 class ProfileFragment : Fragment(R.layout.fragment_profile)
